requirements.txt
openai~=0.2.4
pandas~=1.0.5
dash~=1.13.4
dash-bootstrap-components~=0.10.3
dash-core-components~=1.10.1
dash-html-components~=1.0.3
dash-table~=4.8.1
pathlib2~=2.3.5
python-dotenv~=0.12.0
plotly~=4.4.1